training/requirements.txt
torch>=2.2
torchvision>=0.17
Pillow
numpy
scipy
opencv-python-headless
albumentations
reedsolo
boto3